requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
docker==7.1.0
fastcore==1.8.8
frozenlist==1.7.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
multidict==6.6.4
packaging==25.0
propcache==0.3.2
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyvespa==0.61.0
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
